267.2. Finding Unique User Ids(Section 19: Redux Store Design)
So to make sure that we don't get all these duplicate requests now because we have a single action creator

that fetches all of our data for us we would open up the user hetter action creator.

So here we are right here user hetter.

It attempts to fetch its own data.

So we don't want the user header to fetch its own data anymore.

So we can move that component did mount next.

We don't really need the fetch user action Creator in here at all.

So I would remove that action creator up you're at the top and then we could remove the action creator

list from the connect function down here as well.

Like so.

All right so now we can flip back over.

I still see my list of users loaded up but if I look at my request log I've only made the very bare

minimum number of requests we fetched users ID one all the way through 10 and that's it.

So like I said two possible ways of solving this over fetching issue.

Now when we look at this newly created action creator right here the logic inside of it is pretty darned

compact.

And I think that even though this load ash syntax might be a little bit off I think that eventually

you would be able to look at this thing and understand what was going on pretty easily.

In addition we still have fecche posts as an action creator and fetch user so we can still fetch a list

of posts somewhere else inside of our application if we want it to and we can still fetch an individual

user somewhere else if we wanted to as well.

Finally I think that this approach is a little bit more clear than the memoires approach because like

we said before you would really have to stare at this code for a little bit to understand what was really

going on with it.

All right.

All right.

So a quick pause right here.

When we come back the next section we're going to do just one tiny little refactor to fetch Posten users.

It is totally optional.

You don't have to do it if you want to.

I just want to show you an alternate way of making these couple little calls with load Aktion side of

your.

So quick break and I'll see you in just a minute.