276.3. How Paths Get Matched(Section 20: Navigation with React Router)
Now in order to override this behavior we can find all of our given routes and we can optionally add

on a prop of exact Lysa to see note when we type out a key word inside of a GSX tag like so it is equivalent

to saying equals true.

So anytime you list out just a property name the component that you added to interpret that as having

a prop equal to true.

So inside of the component it would think that this Propst not exact would equal the value true.

So that's what happens when you just put the key word like that by itself.

When you add on a prop name of exact it causes re-act rider to change the rule that it uses for matching

the power slightly.

And as you would expect it essentially changes it to does extracted path equal the string path which

is way closer to how a lot of people probably imagine this system would or should work.

So now with the key word on there we would say does the route of slash page to exactly equal the string

slash.

Well no it doesn't.

So we would now not show that route component on the screen however does flash page to contain slash

page 2.

Yes it still does.

So we would show that route component on the screen.

Now last thing I want to understand is that when we add on this exact keyword it only modifies the route

that you add it to.

So if we had a second route right here with a path of slash that does not have the exact word then this

one right here would be shown on the screen when the user goes to slash page 2.

And so if I say save that you see that I see Page 1 appear again.

Look at Saddam we have a better idea of how reac router works internally and how it matches this path

in this kind of confusing fashion.

Well let's continue with our application in the next video.